Reduce bandwidth used by game state updates.
Remove full transmission of complete card set with every tick and update.
The full card sets are not used by the client, except to obtain the remaining card counts.
Add remaining black and white card fields to the Game model, as well as associated update mechanism.
Note: There may be a cleaner way to do this beyond storing metadata in the model, but I couldn't figure it out.